requirements.txt
altgraph==0.17.4
annotated-types==0.6.0
auto-py-to-exe==2.42.0
bcrypt==4.0.1
bottle==0.12.25
bottle-websocket==0.2.9
cffi==1.16.0
contourpy==1.1.1
customtkinter==5.2.1
cycler==0.12.1
darkdetect==0.8.0
Eel==0.16.0
fonttools==4.43.1
future==0.18.3
gevent==23.9.1
gevent-websocket==0.10.1
greenlet==3.0.1
kiwisolver==1.4.5
matplotlib==3.8.1
mysql-connector-python==8.2.0
numpy==1.26.1
packaging==23.2
pandas==2.1.2
pefile==2023.2.7
Pillow==10.1.0
protobuf==4.21.12
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32-ctypes==0.2.2
seaborn==0.13.0
six==1.16.0
ttkbootstrap==1.10.1
typing_extensions==4.8.0
tzdata==2023.3
whichcraft==0.6.1
zope.event==5.0
zope.interface==6.1
